+++ internal/crypto/requests.go
@@ -1,9 +1,11 @@
 package crypto
 
 import (
+	"crypto/sha256"
 	"errors"
 	"fmt"
 
+	"github.com/nspcc-dev/neo-go/pkg/crypto/hash"
 	apistatus "github.com/nspcc-dev/neofs-sdk-go/client/status"
 	neofscrypto "github.com/nspcc-dev/neofs-sdk-go/crypto"
 	"github.com/nspcc-dev/neofs-sdk-go/proto/refs"
@@ -14,13 +16,28 @@ import (
 // VerifyRequestSignatures checks whether all request signatures are set and
 // valid. Returns [apistatus.SignatureVerification] otherwise.
 func VerifyRequestSignatures[B neofscrypto.ProtoMessage](req neofscrypto.SignedRequest[B]) error {
-	err := neofscrypto.VerifyRequestWithBuffer(req, nil)
+	return verifyRequestSignatures(req, nil)
+}
+
+func verifyRequestSignatures[B neofscrypto.ProtoMessage](req neofscrypto.SignedRequest[B], verifyN3 func(data, invocScript, verifScript []byte) error) error {
+	err := neofscrypto.VerifyRequestWithBufferN3(req, nil, verifyN3)
 	if err != nil {
 		var st apistatus.SignatureVerification
 		st.SetMessage(err.Error())
 		return st
 	}
 	return nil
+}
+
+// VerifyRequestSignaturesN3 is same as [VerifyRequestSignatures] but supports
+// [neofscrypto.N3] scheme.
+func VerifyRequestSignaturesN3[B neofscrypto.ProtoMessage](req neofscrypto.SignedRequest[B], fsChain N3ScriptRunner) error {
+	return verifyRequestSignatures(req, func(data, invocScript, verifScript []byte) error {
+		verifScriptHash := hash.Hash160(verifScript)
+		return verifyN3ScriptsNow(fsChain, verifScriptHash, invocScript, verifScript, func() [sha256.Size]byte {
+			return sha256.Sum256(data)
+		})
+	})
 }
 
 // GetRequestAuthor returns ID of the request author along with public key from
@@ -49,5 +66,7 @@ func GetRequestAuthor(vh *protosession.RequestVerificationHeader) (user.ID, []by
 		//  Not so big overhead, but still better to avoid this
 		ecdsaPub, _ := decodeECDSAPublicKey(sig.Key)
 		return user.NewFromECDSAPublicKey(*ecdsaPub), sig.Key, nil
+	case refs.SignatureScheme_N3:
+		return user.NewFromScriptHash(hash.Hash160(sig.Key)), sig.Key, nil
 	}
 }